% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatter.R
\name{area}
\alias{area}
\title{Create an area to apply formatter}
\usage{
area(row, col)
}
\arguments{
\item{row}{an expression of row range}

\item{col}{an expression of column range}
}
\description{
Create an area to apply formatter
}
\examples{
area(col = c("mpg", "cyl"))
area(col = mpg:cyl)
area(row = 1)
area(row = 1:10, col = 5:10)
}

