% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.datatable.R
\name{as.datatable.formattable}
\alias{as.datatable.formattable}
\title{Convert \code{formattable} to a \code{\link[DT]{datatable}} htmlwidget}
\usage{
\method{as.datatable}{formattable}(x, escape = FALSE, ...)
}
\arguments{
\item{x}{a \code{formattable} object to convert}

\item{escape}{\code{logical} to escape \code{HTML}. The default is
\code{FALSE} since it is expected that \code{formatters} from
\code{formattable} will produce \code{HTML} tags.}

\item{...}{additional arguments passed to to \code{\link[DT]{datatable}}}
}
\value{
a \code{\link[DT]{datatable}} object
}
\description{
Convert \code{formattable} to a \code{\link[DT]{datatable}} htmlwidget
}

