% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formattable_to_DT.R
\name{as.datatable}
\alias{as.datatable}
\title{Generic function to create a datatable htmlwidget}
\usage{
as.datatable(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{arguments to be passed to \code{\link[DT]{datatable}}}
}
\value{
a \code{\link[DT]{datatable}} object
}
\description{
This function is a generic function to create an \code{htmlwidget}
to allow HTML/JS from R in multiple contexts.
}

