% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formattable_to_DT.R
\name{as.htmldf.formattable}
\alias{as.htmldf.formattable}
\title{Convert a \code{\link{formattable}} object into a \code{data.frame}
 on which \code{formatters} have been applied.  This is an intermediate
 step in conversion of a \code{\link{formattable}} into other forms
 such as a \code{\link[DT]{datatable}} htmlwidget.}
\usage{
\method{as.htmldf}{formattable}(x, format_function = "format_table_dataframe")
}
\arguments{
\item{x}{\code{\link{formattable}} object}

\item{format_function}{\code{string} name of a function to process
and format.  \code{}}
}
\value{
\code{\link{data.frame}}
}
\description{
Convert a \code{\link{formattable}} object into a \code{data.frame}
 on which \code{formatters} have been applied.  This is an intermediate
 step in conversion of a \code{\link{formattable}} into other forms
 such as a \code{\link[DT]{datatable}} htmlwidget.
}

