% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formattable_to_DT.R
\name{as.htmldf}
\alias{as.htmldf}
\title{Generic function to convert a \code{\link{formattable}}
  object into a \code{data.frame} with cells on which
  \code{formatters} have been applied.}
\usage{
as.htmldf(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{additional arguments}
}
\value{
\code{data.frame}
}
\description{
Generic function to convert a \code{\link{formattable}}
  object into a \code{data.frame} with cells on which
  \code{formatters} have been applied.
}

