% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{as.htmlwidget}
\alias{as.htmlwidget}
\title{Generic function to create an htmlwidget}
\usage{
as.htmlwidget(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{arguments to be passed to methods.}
}
\value{
a \code{htmlwidget} object
}
\description{
This function is a generic function to create an \code{htmlwidget}
to allow HTML/JS from R in multiple contexts.
}
