% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatter.R
\name{color_bar}
\alias{color_bar}
\title{Create a color-bar formatter}
\usage{
color_bar(color, fun, ...)
}
\arguments{
\item{color}{the background color of the bars}

\item{fun}{the transform function that maps the input vector to
values from 0 to 1}

\item{...}{additional parameters passed to \code{fun}}
}
\description{
Create a color-bar formatter
}
\examples{
formattable(mtcars, list(mpg = color_bar("lightgray", proportion)))
}
\seealso{
\link{normalize_bar}, \link{proportion_bar}
}

