% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatter.R
\name{color_tile}
\alias{color_tile}
\title{Create a color-tile formatter}
\usage{
color_tile(...)
}
\arguments{
\item{...}{parameters passed to \code{gradient}.}
}
\description{
Create a color-tile formatter
}
\examples{
formattable(mtcars, list(mpg = color_tile("white", "pink")))
}
