% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formats.R
\name{currency}
\alias{currency}
\title{Numeric vector with currency format}
\usage{
currency(x, symbol = "$", digits = 2L, format = "f", big.mark = ",",
  ..., sep = "")
}
\arguments{
\item{x}{a numeric vector.}

\item{symbol}{currency symbol}

\item{digits}{an integer to indicate the number of digits of the percentage string.}

\item{format}{format type passed to \code{\link{formatC}}.}

\item{big.mark}{thousands separator}

\item{...}{additional parameters passed to \code{formattable}.}

\item{sep}{separator between symbol and value}
}
\description{
Numeric vector with currency format
}
\examples{
currency(200000)
currency(200000, "\\U20AC")
currency(1200000, "USD", sep = " ")
currency(1200000, "USD", format = "d", sep = " ")
}

