% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formats.R
\name{datetime}
\alias{datetime}
\title{Date/time vector with formatting}
\usage{
datetime(x, format = "", ...)
}
\arguments{
\item{x}{a Date, POSIXct, or POSIXlt object.}

\item{format}{format type passed to \code{\link{format}}.}

\item{...}{additional parameter passed to \code{formattable}.}
}
\description{
Date/time vector with formatting
}
\examples{
datetime(as.Date("2015-02-01"))
datetime(as.Date("2015-02-01"), format = "\%m/\%d/\%Y")
datetime(Sys.time(), format = "\%Y\%m\%dT\%H\%M\%S")
}

