% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{digits}
\alias{digits}
\title{Numeric vector showing pre-specific digits}
\usage{
digits(x, digits, format = "f", ...)
}
\arguments{
\item{x}{a numeric vector}

\item{digits}{an integer to indicate the number of digits to show.}

\item{format}{format type passed to \code{\link{formatC}}.}

\item{...}{additional parameters passed to \code{formattable}.}
}
\description{
Numeric vector showing pre-specific digits
}
\examples{
digits(pi, 2)
digits(123.45678, 3)
}

