% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formattable_to_DT.R
\name{format_table_dataframe}
\alias{format_table_dataframe}
\title{Convert a \code{link{formattable}} into a \code{data.frame}}
\usage{
format_table_dataframe(x, formatters = list(), align = "r",
  row.names = rownames(x), check.rows = FALSE, check.names = FALSE)
}
\arguments{
\item{x}{\code{link{formattable}} object}

\item{formatters}{named \code{list} of \code{\link{formatters}} to
apply to each column.  These likely have already been specified
and available in the \code{attributes} of the
\code{link{formattable}} object.  If already provided, then this
argument is not necessary.}

\item{align}{\code{string} "r". I believe this
is not necessary and only a remnant of copy/paste, but I
need to test further.}

\item{row.names}{see \code{\link{data.frame}}}

\item{check.rows}{see \code{\link{data.frame}}}

\item{check.names}{see \code{\link{data.frame}}}
}
\description{
Convert a \code{link{formattable}} into a \code{data.frame}
}

