% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formattable.R
\name{formattable.data.frame}
\alias{formattable.data.frame}
\title{Attach column formatters to data frame}
\usage{
\method{formattable}{data.frame}(x, ..., formatter = "format_table",
  preproc = NULL, postproc = NULL)
}
\arguments{
\item{x}{a \code{data.frame}.}

\item{...}{arguments to be passed to \code{format_table}.}
}
\value{
a \code{formattable} object
}
\description{
Attach column formatters to data frame
}
\examples{
# mtcars (mpg in red)
formattable(mtcars,
   list(mpg = formatter("span", style = "color:red")))

# mtcars (mpg in red if greater than median)
formattable(mtcars, list(mpg = formatter("span",
   style = function(x) ifelse(x > median(x), "color:red", NA))))

# mtcars (mpg in red if greater than median, using formula)
formattable(mtcars, list(mpg = formatter("span",
   style = x ~ ifelse(x > median(x), "color:red", NA))))

# mtcars (mpg in gradient: the higher, the redder)
formattable(mtcars, list(mpg = formatter("span",
   style = x ~ style(color = rgb(x/max(x), 0, 0)))))

# mtcars (mpg background in gradient: the higher, the redder)
formattable(mtcars, list(mpg = formatter("span",
   style = x ~ style(display = "block",
   "border-radius" = "4px",
   "padding-right" = "4px",
   color = "white",
   "background-color" = rgb(x/max(x), 0, 0)))))
}

