% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formattable.R
\name{formattable.default}
\alias{formattable.default}
\title{Create a formattable object}
\usage{
\method{formattable}{default}(x, ..., formatter = "formatC", preproc = NULL,
  postproc = NULL)
}
\arguments{
\item{x}{an object.}

\item{...}{arguments to be passed to \code{formatter}.}

\item{formatter}{formatting function, \code{formatC} in default.}

\item{preproc}{pre-processor function that prepares \code{x} for
formatting function.}

\item{postproc}{post-processor function that transforms formatted
output for printing.}
}
\value{
a \code{formattable} object that inherits from the original
object.
}
\description{
Create a formattable object
}
\examples{
formattable(rnorm(10), digits = 1)
formattable(rnorm(10), format = "f", digits = 1)
formattable(rnorm(10), format = "f",
  flag="+", digits = 1)
formattable(1:10,
  postproc = function(str, x) paste0(str, "px"))
formattable(1:10,
  postproc = function(str, x)
    paste(str, ifelse(x <= 1, "unit", "units")))
}

