% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/formattable.R
\name{formattable}
\alias{formattable}
\title{Create formatted table by column formatters}
\usage{
formattable(data, formatter = list(), ...)
}
\arguments{
\item{data}{data}

\item{formatter}{A named list of formatter functions.}

\item{...}{additional parameters passed to \code{knitr::kable}}

\item{format}{The output format: markdown or pandoc?}

\item{align}{The alignment of columns: a character vector consisting of \code{'l'} (left),
\code{'c'} (center), and/or \code{'r'} (right). By default, all columns are right-aligned.}
}
\value{
a \code{knitr_kable} object whose \code{print} method generates a
string-representation of \code{data} formatted by \code{formatter} in specific \code{format}.
}
\description{
This is an table generator that specializes in creating
formatted table presented in a mix of markdown/reStructuredText and
HTML elements. To generate a formatted table, each column of data
frame can be transformed by formatter function.
}
\examples{
# mtcars (mpg in red)
formattable(mtcars,
   list(mpg = formatter("span", style = "color:red")))

# mtcars (mpg in red if greater than median)
formattable(mtcars, list(mpg = formatter("span",
   style = function(x) ifelse(x > median(x), "color:red", NA))))

# mtcars (mpg in red if greater than median, using formula)
formattable(mtcars, list(mpg = formatter("span",
   style = x ~ ifelse(x > median(x), "color:red", NA))))

# mtcars (mpg in gradient: the higher, the redder)
formattable(mtcars, list(mpg = formatter("span",
   style = x ~ style(color = rgb(x/max(x), 0, 0)))))

# mtcars (mpg background in gradient: the higher, the redder)
formattable(mtcars, list(mpg = formatter("span",
   style = x ~ style(display = "block",
   "border-radius" = "4px",
   "padding-right" = "4px",
   color = "white",
   "background-color" = rgb(x/max(x), 0, 0)))))
}

