% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{formattableOutput}
\alias{formattableOutput}
\title{Widget output function for use in Shiny}
\usage{
formattableOutput(outputId, width = "100\%", height = "0")
}
\arguments{
\item{outputId}{output variable to read from}

\item{width}{a valid \code{CSS} width or a number}

\item{height}{valid \code{CSS} height or a number}
}
\description{
Widget output function for use in Shiny
}

