% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old-generics.R
\name{legacy}
\alias{legacy}
\alias{digits}
\alias{scientific}
\alias{accounting}
\alias{comma}
\alias{currency}
\alias{percent}
\title{Legacy generics}
\usage{
digits(x, digits, format = "f", ...)

scientific(x, format = c("e", "E"), ...)

accounting(x, digits = 2L, format = "f", big.mark = ",", ...)

comma(x, digits, format = "f", big.mark = ",", ...)

currency(x, symbol, digits, format = "f", big.mark = ",", ...)

percent(x, digits, format = "f", ...)
}
\arguments{
\item{x}{a numeric vector}

\item{digits}{an integer to indicate the number of digits to show.}

\item{format}{format type passed to \code{\link[=formatC]{formatC()}}.}

\item{...}{additional parameters passed to \code{formattable}.}

\item{big.mark}{thousands separator}

\item{symbol}{currency symbol}
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#superseded'><img src='figures/lifecycle-superseded.svg' alt='Superseded lifecycle'></a>}}{\strong{Superseded}}

These generics and functions are superseded.
New code should use the \code{num_...()} or \code{parse_...()} functions instead,
such as \code{\link[=num_accounting]{num_accounting()}}.
}
\keyword{internal}
