% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transform.R
\name{normalize}
\alias{normalize}
\title{Normalize a vector to fit zero-to-one scale}
\usage{
normalize(x, na.rm = FALSE)
}
\arguments{
\item{x}{a vector}

\item{na.rm}{a logical indicating whether missing values
should be removed}
}
\description{
Normalize a vector to fit zero-to-one scale
}
\examples{
normalize(mtcars$mpg)
}

