% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatter.R
\name{normalize_bar}
\alias{normalize_bar}
\title{Create a color-bar formatter using normalize}
\usage{
normalize_bar(color = "lightgray", ...)
}
\arguments{
\item{color}{the background color of the bars}

\item{...}{additional parameters passed to \code{\link[=normalize]{normalize()}}}
}
\description{
Create a color-bar formatter using normalize
}
\examples{
\dontshow{if (requireNamespace("htmlwidgets", quietly = TRUE)) withAutoprint(\{ # examplesIf}
formattable(mtcars, list(mpg = normalize_bar()))
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=color_bar]{color_bar()}}, \code{\link[=normalize]{normalize()}}
}
