% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{proportion}
\alias{proportion}
\title{Rescale a vector relative to the maximal absolute value in the vector}
\usage{
proportion(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector}

\item{na.rm}{a logical indicating whether missing values should be removed}
}
\description{
Rescale a vector relative to the maximal absolute value in the vector
}
\examples{
\dontshow{if (requireNamespace("htmlwidgets", quietly = TRUE)) withAutoprint(\{ # examplesIf}
proportion(mtcars$mpg)
\dontshow{\}) # examplesIf}
}
