% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatter.R
\name{proportion_bar}
\alias{proportion_bar}
\title{Create a color-bar formatter using proportion}
\usage{
proportion_bar(color = "lightgray", ...)
}
\arguments{
\item{color}{the background color of the bars}

\item{...}{additional parameters passed to \code{proportion}}
}
\description{
Create a color-bar formatter using proportion
}
\examples{
formattable(mtcars, list(mpg = proportion_bar()))
}
\seealso{
\link{color_bar}, \link{proportion}
}

