% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{qrank}
\alias{qrank}
\title{Quantile ranks of a vector}
\usage{
qrank(x, ...)
}
\arguments{
\item{x}{a vector}

\item{...}{additional parameters passed to \code{rank}}
}
\description{
The quantile rank of a number in a vector is the relative
position of ranking resulted from rank divided by the length
of vector.
}
\examples{
qrank(mtcars$mpg)
}
\seealso{
\code{\link{rank}}
}
