% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{renderFormattable}
\alias{renderFormattable}
\title{Widget render function for use in Shiny}
\usage{
renderFormattable(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{an expression that generates a valid \code{formattable} object}

\item{env}{the environment in which to evaluate expr.}

\item{quoted}{is expr a quoted expression (with quote())?
This is useful if you want to save an expression in a variable.}
}
\description{
Widget render function for use in Shiny
}

