% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formats.R
\name{scientific}
\alias{scientific}
\title{Numeric vector with scientific format}
\usage{
scientific(x, format = "e", ...)
}
\arguments{
\item{format}{format type passed to \code{\link{formatC}}.}

\item{...}{additional parameter passed to \code{formattable}.}
}
\description{
Numeric vector with scientific format
}

