% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formatter.R
\name{color_bar}
\alias{color_bar}
\title{Create a color-bar formatter}
\usage{
color_bar(color, ...)
}
\arguments{
\item{color}{the background color of the bars}

\item{...}{parameters passed to \code{normalize} function.}
}
\description{
Create a color-bar formatter
}
\examples{
formattable(mtcars, list(mpg = color_bar("pink", 0.2)))
}

