% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatter.R
\name{color_text}
\alias{color_text}
\title{Create a color-text formatter}
\usage{
color_text(...)
}
\arguments{
\item{...}{parameters passed to \code{\link[=gradient]{gradient()}}.}
}
\description{
Create a color-text formatter
}
\examples{
\dontshow{if (requireNamespace("htmlwidgets", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
formattable(mtcars, list(mpg = color_text("black", "red")))
\dontshow{\}) # examplesIf}
}
