% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formats.R
\name{currency}
\alias{currency}
\title{Numeric vector with currency format}
\usage{
currency(x, symbol = "$", digits = 2L, format = "f", big.mark = ",",
  ...)
}
\arguments{
\item{x}{a numeric vector.}

\item{symbol}{currency symbol}

\item{digits}{an integer to indicate the number of digits of the percentage string.}

\item{format}{format type passed to \code{\link{formatC}}.}

\item{big.mark}{thousands separator}

\item{...}{additional parameters passed to \code{formattable}.}
}
\description{
Numeric vector with currency format
}

