% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formattable.R
\name{formattable.default}
\alias{formattable.default}
\title{Create a formattable object}
\usage{
\method{formattable}{default}(x, ..., formatter, preproc = NULL, postproc = NULL)
}
\arguments{
\item{x}{an object.}

\item{...}{arguments to be passed to \code{formatter}.}

\item{formatter}{formatting function, \code{\link[=formatC]{formatC()}} in default.}

\item{preproc}{pre-processor function that prepares \code{x} for
formatting function.}

\item{postproc}{post-processor function that transforms formatted
output for printing.}
}
\value{
a \code{formattable} object that inherits from the original
object.
}
\description{
Create a formattable object
}
\examples{
formattable(rnorm(10), formatter = "formatC", digits = 1)
}
