% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formattable.R
\name{formattable}
\alias{formattable}
\title{Generic function to create formattable object}
\usage{
formattable(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{arguments to be passed to methods.}
}
\value{
a \code{formattable} object
}
\description{
This function is a generic function to create \code{formattable}
object, i.e. an object to which a formatting function and
related attribute are attached. The object works as ordinary object
yet has specially defined behavior as being printed or converted to
a string representation.
}

