% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formatter.R
\name{formatter}
\alias{formatter}
\title{Create a formatter function making HTML elements}
\usage{
formatter(.tag, ...)
}
\arguments{
\item{.tag}{HTML tag name}

\item{...}{functions to create attributes of HTML element from data colums.
The unnamed element will serve as the function to produce the inner text of the
element. If no unnamed element is provided, \code{identity} function will be used
to preserve the string representation of the colum values.}
}
\value{
a function that transforms a column of data (usually an atomic vector)
to formatted data represented in HTML and CSS.
}
\description{
Create a formatter function making HTML elements
}
\examples{
top10red <- formatter("span",
  style = x ~ ifelse(rank(-x) <= 10, "color:red", NA))
yesno <- function(x) ifelse(x, "yes", "no")
formattable(mtcars, list(mpg = top10red, qsec = top10red, am = yesno))
}

