% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formattable.R
\name{is.formattable}
\alias{is.formattable}
\title{Test for objects of 'formattable' class}
\usage{
is.formattable(x)
}
\arguments{
\item{x}{an object}
}
\value{
\code{TRUE} if \code{x} has class 'formattable';
\code{FALSE} otherwise.
}
\description{
Test for objects of 'formattable' class
}
\examples{
is.formattable(10)
is.formattable(formattable(10))
}

