% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old-generics.R
\name{legacy}
\alias{legacy}
\alias{accounting}
\alias{comma}
\title{Legacy generics}
\usage{
accounting(x, digits = 2L, format = "f", big.mark = ",", ...)

comma(x, digits, format = "f", big.mark = ",", ...)
}
\arguments{
\item{x}{a numeric vector.}

\item{digits}{an integer to indicate the number of digits of the percentage string.}

\item{format}{format type passed to \code{\link[=formatC]{formatC()}}.}

\item{big.mark}{thousands separator}

\item{...}{additional parameters passed to \code{formattable}.}
}
\description{
\ifelse{html}{\figure{lifecycle-deprecated.svg}{options: alt='Deprecated lifecycle'}}{\strong{Deprecated}}

These generics are soft-deprecated.
New code should use the \code{num_...()} or \code{parse_...()} functions instead,
such as \code{\link[=num_accounting]{num_accounting()}}.
}
\keyword{internal}
