% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{normalize}
\alias{normalize}
\title{Normalize a vector to fit zero-to-one scale}
\usage{
normalize(x, min = 0, max = 1, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector}

\item{min}{numeric value. The lower bound of the interval to normalize \code{x}.}

\item{max}{numeric value. The upper bound of the interval to normalize \code{x}.}

\item{na.rm}{a logical indicating whether missing values should be removed}
}
\description{
Normalize a vector to fit zero-to-one scale
}
\examples{
normalize(mtcars$mpg)
}

