% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatter.R
\name{normalize_bar}
\alias{normalize_bar}
\title{Create a color-bar formatter using normalize}
\usage{
normalize_bar(color = "lightgray", ...)
}
\arguments{
\item{color}{the background color of the bars}

\item{...}{additional parameters passed to \code{normalize}}
}
\description{
Create a color-bar formatter using normalize
}
\examples{
formattable(mtcars, list(mpg = normalize_bar()))
}
\seealso{
\link{color_bar}, \link{normalize}
}
