% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_accounting.R
\name{num_accounting}
\alias{num_accounting}
\alias{parse_accounting}
\title{Numeric vector with accounting format}
\usage{
num_accounting(x, digits = 2L, format = "f", big.mark = ",", ...)

parse_accounting(
  x,
  digits = max(get_digits(x)),
  format = "f",
  big.mark = ",",
  ...
)
}
\arguments{
\item{x}{a numeric vector}

\item{digits}{an integer to indicate the number of digits to show.}

\item{format}{format type passed to \code{\link[=formatC]{formatC()}}.}

\item{big.mark}{thousands separator}

\item{...}{additional parameters passed to \code{formattable}.}
}
\description{
Formats numbers with a thousand separator
and two digits after the decimal point.
}
\examples{
num_accounting(15320)
num_accounting(-12500)
num_accounting(c(1200, -3500, 2600), format = "d")
parse_accounting(c("123,23.50", "(123.243)"))
}
\seealso{
Other numeric vectors: 
\code{\link{num_comma}()},
\code{\link{num_currency}()},
\code{\link{num_digits}()},
\code{\link{num_percent}()},
\code{\link{num_scientific}()}
}
\concept{numeric vectors}
