% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_comma.R
\name{num_comma}
\alias{num_comma}
\alias{parse_comma}
\title{Numeric vector with thousands separators}
\usage{
num_comma(x, digits = 2L, format = "f", big.mark = ",", ...)

parse_comma(x, digits = max(get_digits(x)), format = "f", big.mark = ",", ...)
}
\arguments{
\item{x}{a numeric vector.}

\item{digits}{an integer to indicate the number of digits of the percentage string.}

\item{format}{format type passed to \code{\link[=formatC]{formatC()}}.}

\item{big.mark}{thousands separator}

\item{...}{additional parameters passed to \code{\link[=formattable]{formattable()}}.}
}
\description{
Formats numbers with a thousand separator
and a prespecified number of digits after the decimal point.
}
\examples{
num_comma(1000000)
num_comma(c(1250000, 225000))
num_comma(c(1250000, 225000), format = "d")
parse_comma("123,345.123")
}
\seealso{
Other numeric vectors:
\code{\link[=num_accounting]{num_accounting()}},
\code{\link[=num_currency]{num_currency()}},
\code{\link[=num_digits]{num_digits()}},
\code{\link[=num_percent]{num_percent()}},
\code{\link[=num_scientific]{num_scientific()}}
}
\concept{numeric vectors}
