% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_currency.R
\name{num_currency}
\alias{num_currency}
\alias{parse_currency}
\title{Numeric vector with currency format}
\usage{
num_currency(
  x,
  symbol = "$",
  digits = 2L,
  format = "f",
  big.mark = ",",
  ...,
  sep = ""
)

parse_currency(
  x,
  symbol = get_currency_symbol(x),
  digits = max(get_digits(x)),
  format = "f",
  big.mark = ",",
  ...
)
}
\arguments{
\item{symbol}{currency symbol}

\item{sep}{separator between symbol and value}
}
\description{
Formats numbers with a thousand separator
and a currency indicator.
}
\examples{
num_currency(200000)
num_currency(200000, "\U20AC")
num_currency(1200000, "USD", sep = " ")
num_currency(1200000, "USD", format = "d", sep = " ")
parse_currency("$ 120,250.50")
parse_currency("HK$ 120,250.50", symbol = "HK$")
parse_currency("HK$ 120, 250.50")
}
\seealso{
Other numeric vectors: 
\code{\link{num_accounting}()},
\code{\link{num_comma}()},
\code{\link{num_digits}()}
}
\concept{numeric vectors}
