% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_digits.R
\name{num_digits}
\alias{num_digits}
\title{Numeric vector showing pre-specific digits}
\usage{
num_digits(x, digits, format = "f", ...)
}
\arguments{
\item{x}{a numeric vector}

\item{digits}{an integer to indicate the number of digits to show.}

\item{format}{format type passed to \code{\link[=formatC]{formatC()}}.}

\item{...}{additional parameters passed to \code{formattable}.}
}
\description{
Formats numbers with
a prespecified number of digits after the decimal point.
}
\examples{
num_digits(pi, 2)
num_digits(123.45678, 3)
}
\seealso{
Other numeric vectors: 
\code{\link{num_accounting}()},
\code{\link{num_comma}()},
\code{\link{num_currency}()},
\code{\link{num_percent}()},
\code{\link{num_scientific}()}
}
\concept{numeric vectors}
