% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_percent.R
\name{num_percent}
\alias{num_percent}
\alias{parse_percent}
\title{Numeric vector with percentage representation}
\usage{
num_percent(x, digits = 2L, format = "f", ...)

parse_percent(x, digits = NA, format = "f", ...)
}
\arguments{
\item{x}{a numeric vector.}

\item{digits}{an integer to indicate the number of digits of the percentage string.}

\item{format}{format type passed to \code{\link[=formatC]{formatC()}}.}

\item{...}{additional parameters passed to \code{\link[=formattable]{formattable()}}.}
}
\description{
Formats numbers as percentages.
}
\examples{
num_percent(rnorm(10, 0, 0.1))
num_percent(rnorm(10, 0, 0.1), digits = 0)
parse_percent("0.5\%")
parse_percent(c("15.5\%", "25.12\%", "73.5"))
}
\seealso{
Other numeric vectors:
\code{\link[=num_accounting]{num_accounting()}},
\code{\link[=num_comma]{num_comma()}},
\code{\link[=num_currency]{num_currency()}},
\code{\link[=num_digits]{num_digits()}},
\code{\link[=num_scientific]{num_scientific()}}
}
\concept{numeric vectors}
