% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formats.R
\name{percent}
\alias{percent}
\title{Percentage representation of numeric values}
\usage{
percent(x, digits = 2L, format = "f", ...)
}
\arguments{
\item{x}{a numeric vector.}

\item{digits}{an integer to indicate the number of digits of the percentage string.}

\item{format}{format type passed to \code{\link{formatC}}.}

\item{...}{additional parameters passed to \code{formattable}.}
}
\description{
Percentage representation of numeric values
}
\examples{
percent(rnorm(10, 0, 0.1))
percent(rnorm(10, 0, 0.1), digits = 0)
}

