% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatter.R
\name{proportion_bar}
\alias{proportion_bar}
\title{Create a color-bar formatter using proportion}
\usage{
proportion_bar(color = "lightgray", ...)
}
\arguments{
\item{color}{the background color of the bars}

\item{...}{additional parameters passed to \code{\link[=proportion]{proportion()}}}
}
\description{
Create a color-bar formatter using proportion
}
\examples{
\dontshow{if (requireNamespace("htmlwidgets", quietly = TRUE)) withAutoprint(\{ # examplesIf}
formattable(mtcars, list(mpg = proportion_bar()))
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=color_bar]{color_bar()}}, \code{\link[=proportion]{proportion()}}
}
