% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formats.R
\name{scientific}
\alias{scientific}
\title{Numeric vector with scientific format}
\usage{
scientific(x, format = c("e", "E"), ...)
}
\arguments{
\item{x}{a numeric vector.}

\item{format}{format type passed to \code{\link{formatC}}.}

\item{...}{additional parameter passed to \code{formattable}.}
}
\description{
Numeric vector with scientific format
}
\examples{
scientific(1250000)
scientific(1253421, digits = 8)
scientific(1253421, digits = 8, format = "E")
}

